data/requirements.txt
pymupdf==1.24.9
pillow==10.4.0
pytesseract==0.3.13
tqdm==4.67.1
opencv-python==4.10.0.84
transformers==4.44.2
numpy==1.26.4
pandas==2.2.3
pyarrow==18.1.0
spacy==3.7.5
scikit-learn==1.5.2
